Run 'cargo clippy --fix'

+++ src/disk/sys/unix/disk_usage.rs
@@ -44,12 +44,12 @@ where
 	// https://github.com/rust-psutil/rust-psutil/issues/64
 	// https://github.com/rust-psutil/rust-psutil/pull/39
 
-	let total = statvfs.blocks() as u64 * statvfs.fragment_size() as u64;
+	let total = statvfs.blocks() * statvfs.fragment_size();
 
-	let avail_to_root = statvfs.blocks_free() as u64 * statvfs.fragment_size() as u64;
+	let avail_to_root = statvfs.blocks_free() * statvfs.fragment_size();
 	let used = total - avail_to_root;
 
-	let free = statvfs.blocks_available() as u64 * statvfs.fragment_size() as u64;
+	let free = statvfs.blocks_available() * statvfs.fragment_size();
 
 	let total_user = used + free;
 	let percent = ((used as f64 / total_user as f64) * 100.0) as f32;
